Create home module and move Home composable to new module.
