+++ android/app/src/main/kotlin/com/thomaskioko/tvmaniac/RootScreen.kt
@@ -14,9 +14,9 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import com.arkivanov.decompose.extensions.compose.stack.Children
+import com.thomaskioko.tvmaniac.home.HomeContent
 import com.thomaskioko.tvmaniac.navigation.RootComponent
 import com.thomaskioko.tvmaniac.seasondetails.ui.SeasonDetailsScreen
-import com.thomaskioko.tvmaniac.tabs.HomeContent
 import com.thomaskioko.tvmaniac.ui.moreshows.MoreShowsScreen
 import com.thomaskioko.tvmaniac.ui.showdetails.ShowDetailsScreen
 import com.thomaskioko.tvmaniac.ui.trailers.videoplayer.TrailersScreen
