+++ android/ui/home/src/main/java/com/thomaskioko/tvmaniac/home/HomeContent.kt
@@ -1,4 +1,4 @@
-package com.thomaskioko.tvmaniac.tabs
+package com.thomaskioko.tvmaniac.home
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -8,30 +8,27 @@ import androidx.compose.material.icons.outlined.Movie
 import androidx.compose.material.icons.outlined.Search
 import androidx.compose.material.icons.outlined.Settings
 import androidx.compose.material.icons.outlined.VideoLibrary
-import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import com.arkivanov.decompose.extensions.compose.stack.Children
-import com.thomaskioko.tvmaniac.compose.components.ThemePreviews
 import com.thomaskioko.tvmaniac.compose.components.TvManiacBottomNavigationItem
 import com.thomaskioko.tvmaniac.compose.components.TvManiacNavigationBar
-import com.thomaskioko.tvmaniac.compose.theme.TvManiacTheme
 import com.thomaskioko.tvmaniac.presentation.home.HomeComponent
 import com.thomaskioko.tvmaniac.presentation.home.HomeComponent.Child.Discover
 import com.thomaskioko.tvmaniac.presentation.home.HomeComponent.Child.Library
 import com.thomaskioko.tvmaniac.presentation.home.HomeComponent.Child.Search
 import com.thomaskioko.tvmaniac.presentation.home.HomeComponent.Child.Settings
 import com.thomaskioko.tvmaniac.resources.R
-import com.thomaskioko.tvmaniac.search.ui.SearchScreen
+import com.thomaskioko.tvmaniac.ui.search.SearchScreen
 import com.thomaskioko.tvmaniac.ui.discover.DiscoverScreen
 import com.thomaskioko.tvmaniac.ui.library.LibraryScreen
 import com.thomaskioko.tvmaniac.ui.settings.SettingsScreen
 
 @Composable
-internal fun HomeContent(
+fun HomeContent(
   component: HomeComponent,
   modifier: Modifier = Modifier,
 ) {
